+++ application/models/Author_model.php
@@ -77,6 +77,8 @@ class Author_model extends MY_Model
                 'label' => 'User ID',
                 'rules' => 'trim|callback_unique_author_username'
             ]
+            
+            
         ];
 
         return $validationRules;
@@ -128,8 +130,11 @@ class Author_model extends MY_Model
 
     public function deleteAuthorKTP($authorKTP)
     {
-        if (file_exists("./authorktp/$authorKTP")) {
-            unlink("./authorktp/$authorKTP");
+        if($draftFile != "") {
+           if (file_exists("./authorktp/$authorKTP")) {
+                unlink("./authorktp/$authorKTP");
+            } 
         }
+        
     }
 }
